추가 : 약속 나가기 시 채팅방 접속 인원인지 확인하는 api추가, 댓글작성자 친구추가, 포인트 주기, 신고하기 버튼

+++ src/components/header/Option3.jsx
@@ -39,14 +39,23 @@ const Option3 = ({ head }) => {
         }
       }).then(async(response) => {
         if (response.data.success) {
-          const chat = await axios.delete(process.env.REACT_APP_SERVER_HOST+`/api/chat/member/${id}`,{
+          await axios.get(process.env.REACT_APP_SERVER_HOST+`/api/chat/member/${id}`,{
             headers: {
               Authorization: localStorage.getItem('Authorization'),
               RefreshToken: localStorage.getItem('RefreshToken'),
+          }}).then(async(res)=>{
+            if(res.data.data !== "채팅방에 없는 회원입니다"){
+              navigate("/")
+            }else{
+              const chat = await axios.delete(process.env.REACT_APP_SERVER_HOST+`/api/chat/member/${id}`,{
+                headers: {
+                  Authorization: localStorage.getItem('Authorization'),
+                  RefreshToken: localStorage.getItem('RefreshToken'),
+                }
+              })
+              navigate("/")
             }
           })
-          console.log(chat)
-          navigate("/")
         } else {
           alert(response.data.data);
           setChk(0);
